% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSeq.R
\name{DOSeq}
\alias{DOSeq}
\title{Modeling expression drop-out for analysis of scRNA-Seq data}
\usage{
DOSeq(countData, factor = 0.8)
}
\arguments{
\item{countData}{read count matrix, with row names as genes name/ID and
column names as sample id/name}

\item{factor}{indicates dropout ratio}
}
\value{
Simulated read count with drop out
}
\description{
DOSeq - Modeling expression drop-out for analysis of scRNA-Seq
data. DOSeq takes read count matrix and factor for drop-out ratio as input
and return the read count matrix with dropout. Here input parameter factor
indicates drop-out ratio in expression matrix.
}
\examples{
countData<-DOSeq::L_Tung_single$NA19098_NA19101_count
head(countData)
output<-DOSeq(countData=countData, factor=.8)
output
}
